chore: recover changes lost in old repository
When I was migrating the old repository from private to public, for some reason an error occurred that did not allow me to access that repository anymore.

So I created a new repository, and that will count as the initial commit from now on.

The changes I made to the old repository are:

chore: setup expo enviroment
chore: setup API
chore: add fonts && add colors && create basic components